+++ CHANGELOG.md
@@ -1,2 +1,11 @@
-## 1.0.0 - 2018-07-31
-* Initial release with orders, payments and refunds APIs with support for EMI and UPI.
+# Change Log
+
+Changelog for payabbhi-dotnet.
+
+## [1.0.1] - 2018-11-16
+### Added
+- Support for Webhook signature verification
+
+## [1.0.0] - 2018-07-31
+### Added
+- Initial release with orders, payments and refunds APIs with support for EMI and UPI.